requirements.txt
appnope==0.1.2
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
decorator==4.4.2
fake-useragent==0.1.11
httmock==1.4.0
http-request-randomizer==1.3.2
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
ipython==7.19.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
jedi==0.17.2
jmespath==0.10.0
lxml==4.6.2
mysql-connector-python==8.0.22
numpy==1.19.4
pandas==1.1.5
parsel==1.6.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
Protego==0.1.16
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.7.3
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2020.4
queuelib==1.5.0
requests==2.25.1
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
soupsieve==2.1
traitlets==5.0.5
Twisted==20.3.0
urllib3==1.26.2
w3lib==1.22.0
wcwidth==0.2.5
zope.interface==5.2.0
